优化FontHelper/优化FPSProcessor
FPSProcessor的FPS精度已经很高了……

并且FontHelper的FillTexture函数的拷贝Color对象改为拷贝Byte，速度将显著提升……把GameConst中的VSync与LimitFPS关掉就会有效果~

+++ Source/Dawn/Dawn/Engine/Manager/GraphicsManager.cs
@@ -23,7 +23,7 @@ namespace Dawn.Engine.Manager
 		public Microsoft.Xna.Framework.Graphics.GraphicsDevice Device { get { return DGE.Game._GraphicsDevice; } }
 		public Microsoft.Xna.Framework.Graphics.SpriteBatch Canvas { get { return DGE.Game._SpriteBatch; } }
         public SceneManager Scenes { get { return _Scenes; } }
-		public int FPS { get { return fps.FPS; } }
+		public double FPS { get { return fps.FPS; } }
         public GraphicsManager()
         {
             _Scenes=new SceneManager();
